medical_report_analyzer_streamlit/requirements.txt
streamlit==1.31.1
opencv-python==4.8.1.78
pytesseract==0.3.10
pypdf2==3.0.1
pdf2image==1.16.3
numpy==1.26.2
pillow==10.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
reportlab==4.0.9  # For PDF generation 